Issue #4220: Modified EmptyBlockCheckTest.java and moved its input files to the emptyblock subdirectory
